Add GET endpoint to retrieve list of Products from API
This commit adds a new GET endpoint to the API that returns a list of Product objects. The endpoint is accessed by the Blazor eCommerce app to fetch the products for display in the user interface. This feature enhances the app's functionality by providing real-time product data to users.

Episode #16